Replaced labels, preview slider and spinboxes with ctkSliderWidget and adjusted stepsize accordingly

+++ Modules/SegmentationUI/Qmitk/QmitkAdaptiveRegionGrowingToolGUI.h
@@ -86,7 +86,7 @@ protected slots:
 
     void RunSegmentation();
 
-    void ChangeLevelWindow(int newValue);//called, when the Level Window is changed via the slider in the ControlWidget
+    void ChangeLevelWindow(double newValue);//called, when the Level Window is changed via the slider in the ControlWidget
 
     //****called, when the slider-position is modified via the +/- buttons
     void IncreaseSlider();
